refactor: extract GeoServerImportService partials by concern (#1144)
Main file 3,414 → 802 LOC (-76%). Logic grouped into:
- GeoServerImportService.Apply.cs (apply pipeline orchestration)
- GeoServerImportService.Identifiers.cs (id resolution helpers)
- GeoServerImportService.ImportSteps.cs (per-step import operations)
- GeoServerImportService.Inventory.cs (workspace/layer/coverage inventory)

Build clean, no API changes.
